requirements.txt
PySide6
pyserial
gpsd-py3